Moved three js dependency from dependencies
Having an peer dependency in dependecies is error prone since it can
introduce two different versions simultaneously.
For convenience reasons I duplicated it in dev depencdencies.
That way when developing locally you'll get it installed as dev dep
and users of the library will only have a peer dep
This a potencially a breaking change - since users may be not having
three as a direct dependency